+++ CHANGELOG.md
@@ -9,11 +9,15 @@
 - Added support for query `PeerList`
 - Added support for query `PeerStats`
 - Added support for query `PeerVersion`
+- Added support for query `GetModuleSource`
+- Added support for query `GetModuleList`
+- Added support for query `GetBirkParameters`
 - Added support for query `NodeInfo`
 - Added support for query `Shutdown`
 - Added support for query `GetAncestors`
 - Added support for query `JoinNetwork`
 - Added support for query `LeaveNetwork`
+- Added support for query `GetRewardStatus`
 - Added support for query `GetInstanceInfo`
 - Added support for query `GetInstances`
 - Added support for query `GetAccountList`